+++ Demo/Shared/Gestures/GesturesScreen.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2020 Daniel Saidi. All rights reserved.
 //
 
-#if os(iOS)
 import SwiftUI
 
 struct GesturesScreen: View {
@@ -29,4 +28,3 @@ struct GesturesScreen_Previews: PreviewProvider {
         GesturesScreen()
     }
 }
-#endif