+++ Tests/SwiftUIKitTests/Extension/UIKit/UIImage+TintedTests.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
+#if os(iOS)
 import Quick
 import Nimble
 import SwiftUIKit
@@ -25,3 +26,4 @@ class UIImage_TintedTests: QuickSpec {
         }
     }
 }
+#endif